Revert "fix warnings from Intel compiler"
This reverts commit 8b9338ab57eec0cc8fa5a36c44d7b28e52e9a466.

+++ src/H5FAcache.c
@@ -420,7 +420,7 @@ H5FA__cache_hdr_serialize(const H5F_t *f, void *_image, size_t H5_ATTR_UNUSED le
     *image++ = H5FA_HDR_VERSION;
 
     /* Fixed array type */
-    *image++ = (uint8_t)hdr->cparam.cls->id;
+    *image++ = hdr->cparam.cls->id;
 
     /* General array creation/configuration information */
     *image++ = hdr->cparam.raw_elmt_size;          /* Element size in file (in bytes) */
@@ -808,7 +808,7 @@ H5FA__cache_dblock_serialize(const H5F_t *f, void *_image, size_t H5_ATTR_UNUSED
     *image++ = H5FA_DBLOCK_VERSION;
 
     /* Fixed array type */
-    *image++ = (uint8_t)dblock->hdr->cparam.cls->id;
+    *image++ = dblock->hdr->cparam.cls->id;
 
     /* Address of array header for array which owns this block */
     H5F_addr_encode(f, &image, dblock->hdr->addr);
